soc: qcom: rpm-smd: Avoid deadlocking while waiting for ACK from RPM
The RPM may give back no ACK when sending a command: this
means that it is locking up for some reason, or the
command itself didn't get sent to it.

When we get in this situation, an infinite wait for completion
will lock up the device for an infinite amount of time.
Add a 5-second completion timeout to prevent indefinitely deadlocking.